+++ inst/extdata/sql-queries/v2-os-municipios-clean-csv-view-es.sql
@@ -1,7 +1,7 @@
 CREATE OR REPLACE VIEW os_csv_clean AS SELECT
     fecha,
-    CAST(zona_residencia AS RESID_ZONES_ENUM) AS zona_residencia,
-    CAST(zona_pernoctacion AS OVERNIGHT_ZONES_ENUM) AS zona_pernoctacion,
+    CAST (CASE zona_residencia WHEN 'NA' THEN NULL ELSE zona_residencia END AS RESID_ZONES_ENUM) AS zona_residencia,
+    CAST (CASE zona_pernoctacion WHEN 'NA' THEN NULL ELSE zona_pernoctacion END AS OVERNIGHT_ZONES_ENUM) AS zona_pernoctacion,
     personas,
     CAST(year AS INTEGER) AS ano,
     CAST(month AS INTEGER) AS mes,